📄 Update README.md
The README.md file wasn't easily accessible through the Xcode project. I added a file reference to it in the project.

The project description needs to be more descriptive. I added information about the availability and the frameworks used. The link to the associated article was supplemented but kept hidden for now since it isn't released yet.