⭐️ Add new demo project and restructure repository
The repository was initially only supposed to house the "Demo-ShazamKit" project, but it would have been a hassle to have a repository for each demo project I make. I restructured it to make it more flexible and accommodate new projects in the future, including the new "Animate UICollectionViewCell highlight" playground in the "Swift Tips" directory.

Having the prefix "Demo-" in front of every project name was unuseful. I renamed "Demo-ShazamKit" to its future associated article "ShazamKit - A first look". I also put it in the "Swift" directory.

The repository's README.md needs to include a link to every project, organised by type, to make it easily accessible to users. I added these links and associated them with a date. That date would either be the date of publication or, when available, the date of publication of their corresponding article.

Every project needs its README.md file to disclose corresponding articles, availability and frameworks used. I added the file to the two projects with the link to the article commented out before it is published. Projects in the "Swift Tips" category also have a preview video in that file.