🧹 Remove projects and .gitattributes and update README.md
The projects included in the repository aren't individually downloadable.
Therefore, I made a new organisation, and, in it, I added a repository for each of my demo projects, removing them from this repository.

The .gitattributes file doesn't feel necessary.
Therefore, I removed it.

The short description in the README.md isn't descriptive enough, and the links are now inaccurate.
Therefore, I updated the README.md file with a slightly adjusted description and links to the new project repositories in the organisation.